Abstract scatter and line examples into showComponent function
