+++ app/examples2/index.js
@@ -3,4 +3,4 @@ import examples from './index.json';
 
 window.onload = () => {
   document.body.innerHTML = html({examples});
-}
+};
