--- app/examples2/scatter/index.jade
@@ -1,2 +1,0 @@
-.page
-  #vis-container.vis-full
