+++ app/examples2/util/showComponent/index.jade
