+++ app/examples2/util/showComponent/index.js
@@ -1,0 +1,14 @@
+import candela from '../../../../src/candela';
+import html from './index.jade';
+
+export default function showComponent (which, elementType, options) {
+  document.body.innerHTML = html();
+
+  let el = document.getElementById('vis-container')
+  .appendChild(document.createElement(elementType));
+  el.setAttribute('id', 'vis-element');
+  el.className = 'vis-full';
+
+  let vis = new candela.components[which](el, options);
+  vis.render();
+}
